posted chart

+++ 7.expense-tracker/src/components/DeleteButton.jsx
@@ -2,24 +2,25 @@
 import { DeleteIcon } from "@/svg/DeleteIcon";
 
 export const DeleteButton = ({ onDelete, expense }) => {
-  // const isConfirmed = window.confirm(
-  //   "Are you sure you want to delete this expense?"
-  // );
-
   const handleDelete = async () => {
-    const res = await fetch(
-      `http://127.0.0.1:3001/api/expenses/${expense.pk_id}`,
-      {
-        method: "DELETE",
-      }
+    const isConfirmed = window.confirm(
+      "Are you sure you want to delete this expense?"
     );
-    console.log(expense);
+    if (isConfirmed) {
+      const res = await fetch(
+        `http://127.0.0.1:3001/api/expenses/${expense.pk_id}`,
+        {
+          method: "DELETE",
+        }
+      );
+      console.log(expense);
 
-    if (res.ok) {
-      prompt("ARE YOU SURE WANTS TO DELETE EXPENSE");
-      onDelete();
-    } else {
-      console.error("Failed to delete the expense.");
+      if (res.ok) {
+        prompt("ARE YOU SURE WANTS TO DELETE EXPENSE");
+        onDelete();
+      } else {
+        console.error("Failed to delete the expense.");
+      }
     }
   };
